Fix wpt/css/CSS2/floats-clear/clear-on-parent-with-margins-no-clearance.html
The test was incorrect, which is why it failed in legacy layout. It
passed in LayoutNG because of a bug there.

Add an (invisible) top border on the outer DIV to prevent clearance on
the child. Otherwise the outer DIV and the child with 'clear' would have
adjacent margins, which in turn would enable the child with 'clear' to
pull the float along with itself, which in turn would guarantee
clearance, which is precisely what this test wanted to avoid.

Still passes in LayoutNG.

Change-Id: I6b36d8196b93720453d6e398509705d2cf32f8e4